improved code readability and modularity
