+++ frontend/src/data/BASE_API.js
@@ -1,0 +1,1 @@
+export const API_BASE = 'https://api.terraquakeapi.com';
